refactor: move timer to utils
